feat: implement CLI, locking, timestamp formatting, systemd integration
Full implementation of the qsnap 1.0 user-facing layer:

CLI (qsnap/cli/):
- argparse-based CLI with subcommands: run/snapshot/backup/prune/list/stats/check
- Global flags: --config, --dry-run, --preserve, --verbose/quiet, --format, --lockfile
- Output formatters: table (default), long, raw, col:<columns>
- Exit codes: 0=success, 1=generic, 2=parse, 3=lockfile, 10=backup abort
- Thin layer: commands.py has zero business logic imports

Core (qsnap/core/):
- preserve_snapshots/preserve_backups properties (skip deletion on demand)
- 6 informational methods: list_snapshots, list_backups, list_config, list_latest, print_schedule, check
- CheckResult dataclass for backing-chain integrity checks
- Timestamp format resolution via qsnap/utils/time.py

Locking (qsnap/locking.py):
- LockManager with fcntl.flock (non-blocking)
- resolve_lockfile_path: CLI > config > null (no locking)
- No-op mode when lockfile is None

Timestamp formatting (qsnap/utils/time.py):
- resolve_format() and format_snapshot_timestamp() utilities
- Supported: short (%Y%m%d), long (%Y%m%dT%H%M, default), long-iso (%Y%m%dT%H%M%S%z)
- Collision suffix _N for duplicate timestamps

Retention:
- preserve_day_of_week parameter in IRetentionEngine.evaluate()
- Weekly bucket grouping based on configured day (case-insensitive)
- ConfigFacade validation for preserve_day_of_week (monday..sunday)

Configuration:
- GlobalConfig.timestamp_format default changed: short → long
- GlobalConfig.lockfile consumed by LockManager
- GlobalConfig.preserve_day_of_week consumed + validated

Systemd:
- qsnap.service (oneshot, ExecStart=qsnap run)
- qsnap.timer (OnCalendar=hourly, Persistent=true, RandomizedDelaySec=300)
- qsnap.toml.example with full commented documentation

Testing:
- 229 tests (120 baseline + 109 new), all passing
- New: tests/cli/ (test_app, test_commands, test_format, test_thin_layer)
- New: tests/utils/ (test_locking, test_time)
- New: tests/core/ (test_list_commands, test_preserve)
- New: tests/systemd/test_units.py
- Modified: 8 existing test files with new assertions
- ruff: All checks passed, black: All files formatted

+++ qsnap/models/config.py
@@ -36,7 +36,7 @@ class GlobalConfig:
     ``"24h 2d"``) resolved by ``ConfigFacade`` via option inheritance.
     """
 
-    timestamp_format: str = "short"
+    timestamp_format: str = "long"
     preserve_day_of_week: str = "monday"
     state_dir: str = "/var/lib/qsnap/state"
     lockfile: str | None = None
